Hakyll.Core.Identifier

Description

An identifier is a type used to uniquely identify an item. An identifier is conceptually similar to a file path. Examples of identifiers are:

posts/foo.markdown

index

error/404

Documentation

data Identifier

fromFilePath :: String -> Identifier

Parse an identifier from a string

toFilePath :: Identifier -> FilePath

Convert an identifier to a relative FilePath

identifierVersion :: Identifier -> Maybe String

setVersion :: Maybe String -> Identifier -> Identifier