Hakyll.Core.Item

Description

An item is a combination of some content and its Identifier. This way, we can still use the Identifier to access metadata.

Documentation

data Item a

Constructors

Item

Fields

itemIdentifier :: Identifier

itemBody :: a

itemSetBody :: a -> Item b -> Item a

withItemBody :: (a -> Compiler b) -> Item a -> Compiler (Item b)

Perform a compiler action on the item body. This is the same as traverse, but looks less intimidating.

withItemBody = traverse