Hakyll.Main

Description

Module providing the main hakyll function and command-line argument parsing

Documentation

hakyll :: Rules a -> IO ()

This usualy is the function with which the user runs the hakyll compiler

hakyllWith :: Configuration -> Rules a -> IO ()

A variant of hakyll which allows the user to specify a custom configuration