Hakyll.Web.CompressCss

Description

Module used for CSS compression. The compression is currently in a simple state, but would typically reduce the number of bytes by about 25%.

Documentation

compressCssCompiler :: Compiler (Item String)

Compiler form of compressCss

compressCss :: String -> String

Compress CSS to speed up your site.