Hakyll.Web.Feed

Description

A Module that allows easy rendering of RSS feeds.

The main rendering functions (renderRss, renderAtom) all assume that you pass the list of items so that the most recent entry in the feed is the first item in the list.

Also note that the context should have (at least) the following fields to produce a correct feed:

$title$: Title of the item

$description$: Description to appear in the feed

$url$: URL to the item - this is usually set automatically.

In addition, the posts should be named according to the rules for dateField

Documentation

data FeedConfiguration

This is a data structure to keep the configuration of a feed.

Constructors

FeedConfiguration

Fields

feedTitle :: String

Title of the feed.

feedDescription :: String

Description of the feed.

feedAuthorName :: String

Name of the feed author.

feedAuthorEmail :: String

Email of the feed author.

feedRoot :: String

Absolute root URL of the feed site (e.g. http://jaspervdj.be)

renderRss

Arguments

:: FeedConfiguration
Feed configuration

-> Context String
Item context

-> [Item String]
Feed items

-> Compiler (Item String)
Resulting feed

Render an RSS feed with a number of items.

renderAtom

Arguments

:: FeedConfiguration
Feed configuration

-> Context String
Item context

-> [Item String]
Feed items

-> Compiler (Item String)
Resulting feed

Render an Atom feed with a number of items.