Hakyll.Web.Pandoc.Biblio

Description

Wraps pandocs bibiliography handling

In order to add a bibliography, you will need a bibliography file (e.g. .bib) and a CSL file (.csl). Both need to be compiled with their respective compilers (biblioCompiler and cslCompiler). Then, you can refer to these files when you use pageReadPandocBiblio. This function also takes the reader options for completeness -- you can use defaultHakyllReaderOptions if you're unsure.

Documentation

data CSL

cslCompiler :: Compiler (Item CSL)

newtype Biblio

Constructors

Biblio [Reference]

biblioCompiler :: Compiler (Item Biblio)

readPandocBiblio :: ReaderOptions -> Item CSL -> Item Biblio -> Item String -> Compiler (Item Pandoc)