Hakyll.Web.Pandoc.FileType

Description

A module dealing with pandoc file extensions and associated file types

Documentation

data FileType

Datatype to represent the different file types Hakyll can deal with by default

Constructors

Binary

Css

Html

LaTeX

LiterateHaskell FileType

Markdown

OrgMode

PlainText

Rst

Textile

fileType :: FilePath -> FileType

Get the file type for a certain file. The type is determined by extension.

itemFileType :: Item a -> FileType

Get the file type for the current file