Hakyll.Web.Template.List

Description

Provides an easy way to combine several items in a list. The applications are obvious:

A post list on a blog

An image list in a gallery

A sitemap

Documentation

applyTemplateList :: Template -> Context a -> [Item a] -> Compiler String

Generate a string of a listing of pages, after applying a template to each page.

applyJoinTemplateList :: String -> Template -> Context a -> [Item a] -> Compiler String

Join a listing of pages with a string in between, after applying a template to each page.

chronological :: MonadMetadata m => [Item a] -> m [Item a]

Sort pages chronologically. Uses the same method as dateField for extracting the date.

recentFirst :: MonadMetadata m => [Item a] -> m [Item a]

The reverse of chronological

sortChronological :: MonadMetadata m => [Identifier] -> m [Identifier]

Version of chronological which doesn't need the actual items.

sortRecentFirst :: MonadMetadata m => [Identifier] -> m [Identifier]

Version of recentFirst which doesn't need the actual items.